v1.8-SE

+++ bootloader/storage/emummc.h
@@ -17,8 +17,7 @@
 #ifndef EMUMMC_H
 #define EMUMMC_H
 
-#include <storage/sdmmc.h>
-#include <utils/types.h>
+#include <bdk.h>
 
 typedef enum
 {
@@ -37,7 +36,7 @@ typedef struct _emummc_cfg_t
 {
 	int   enabled;
 	u64   sector;
-	u16   id;
+	u32   id;
 	char *path;
 	char *nintendo_path;
 	// Internal.
